Module Equinoxe_hlc.Api

It is the signature that matches the API of the website.

type t

Abstract type t represents the information known by the API system.

val create : ?address:string -> ?token:string -> unit -> t

create ~address ~token () returns an t object, you need to manipulate when executing requests. Default address is https://api.equinix.com/metal/v1/ and default token is empty.

module Auth : sig ... end

This module manages API parts related to authentification.

module Devices : sig ... end

This module manages API parts related to devices.

module Ip : sig ... end

This module manages API parts related to ips.

module Orga : sig ... end

This module manages API parts related to organizations.

module Projects : sig ... end

This module manages API parts related to projects.

module Users : sig ... end

This module manages API parts related to users.