Module Equinoxe_hlc.State

This module represents the state of a Device in the Equinix API.

type t =

| Active

| Queued

| Provisioning

| Inactive

| Powering_off

| Powering_on

Available state to describe a machine.

val of_string : string -> t

of_string str returns a state in function of a string. If the state is not known, it raises an Unknown_value exception.

val to_string : t -> string

to_string t returns a string representation of the state.